Use float SNR and add safe_append helper
Change NodeRecord.snr from int8_t to float and update UI formatting to display SNR with one decimal place. Introduce safe_append(buffer, size, append) to centralize buffer overflow protection and replace repeated manual checks when appending log and chat messages. Ensure long_name copies use a 39-byte strncpy with explicit null termination to avoid overruns when storing node names. These changes preserve fractional SNR precision, reduce risk of buffer overflow, and make node/name handling more robust.

+++ types.h
@@ -170,7 +170,7 @@ struct NodeRecord {
   uint32_t num;
   char long_name[40];
   uint32_t last_heard;
-  int8_t snr;
+  float snr;
 };
 
 // Define LoRa/Serial State
